feat: initial commit
Initial commit to this repository pushed to GitHub for safekeeping and
continuing work on other devices.

For future contributors (including future me) please install the
following:

For developers:
- Godots
  - Godot 4.4+
- VS Code

For artists:
- Blender
- Your preferred image editing suite

This so far has been the bare minimum needed software while working on
this game. Also use the following link as a guide on how this project is
roughly structured so far

https://github.com/abmarnie/godot-architecture-organization-advice